LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCDTMFSender-insertDTMF.https-expected.txt


PASS insertDTMF() should succeed if tones contains valid DTMF characters
PASS insertDTMF() should throw InvalidCharacterError if tones contains invalid DTMF characters
PASS insertDTMF() should throw InvalidStateError if transceiver is stopped
PASS insertDTMF() should throw InvalidStateError if transceiver.currentDirection is recvonly
PASS insertDTMF() should throw InvalidStateError if transceiver.currentDirection is inactive
PASS insertDTMF() should set toneBuffer to provided tones normalized, with old tones overridden
PASS insertDTMF() after remove and close should reject
